made several modifications to the PluginController and PluginService classes:
The POST method for creating a review initially used versionNumber as a parameter, which caused a bug. It has now been replaced with pluginVersionId.

The logic for creating and retrieving PluginVersions and Reviews has been moved from the PluginService class to the PluginVersionService and ReviewService classes, respectively.

The GET and POST methods for PluginVersion and Review have been relocated from PluginController to PluginVersionController and ReviewController, respectively.